helix-backend/requirements.txt
psycopg[binary]>=3.2.0
psycopg-pool>=3.2.0
psycopg2-binary>=2.9.9
langgraph>=0.0.20
langgraph-checkpoint-postgres>=2.0.0
Flask-SQLAlchemy>=3.1.1
Flask-CORS>=4.0.0
Flask-SocketIO>=5.3.6
langchain-google-genai>=0.0.6
langchain-core>=0.1.22
langchain-community>=0.0.24
python-dotenv>=1.0.0
tavily-python>=0.3.1
duckduckgo-search>=4.4.3
SQLAlchemy>=2.0.0
Werkzeug>=3.0.0
Flask>=3.0.0